# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: ISMS Builder API
  version: 1.0.0
  description: |
    REST API for the ISMS Builder platform — a self-hosted Information Security
    Management System (ISMS) according to ISO 27001:2022, NIS2, GDPR, BSI
    IT-Grundschutz and other frameworks.

    **Authentication:** All protected endpoints require a valid JWT cookie
    (`sm_session`). Obtain it via `POST /login`.

    **RBAC Roles (ascending rank):**
    - `reader` / `revision` (rank 1) — read access
    - `editor` / `dept_head` / `qmb` (rank 2) — create/edit
    - `contentowner` / `auditor` (rank 3) — approve, CISO/DSB settings
    - `admin` (rank 4) — user management, delete, system config

  contact:
    name: Claude Hecker
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Self-hosted instance
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sm_session
  schemas:
    CommonFields:
      type: object
      properties:
        id:
          type: string
          example: lx3k8fab2c
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        linkedControls:
          type: array
          items:
            type: string
          example: ["ISO-5.1", "BSI-ORP.3"]
        linkedPolicies:
          type: array
          items:
            type: string
...
    Risk:
      allOf:
        - $ref: '#/components/schemas/CommonFields'
        - type: object
          properties:
            title:
              type: string
            category:
              type: string
            description:
              type: string
            probability:
              type: integer
              minimum: 1
              maximum: 5
            impact:
              type: integer
              minimum: 1
              maximum: 5
            score:
              type: integer
            treatment:
              type: string
              enum: [mitigate, accept, avoid, transfer]
            status:
              type: string
              enum: [open, in_progress, closed]
            owner:
              type: string
            treatments:
              type: array
              items:
                type: object
...
paths:
...
  /risks:
    get:
      tags: [risks]
      summary: List risks
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Risk'
    post:
      tags: [risks]
      summary: Create risk (auditor/contentowner+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Risk'
      responses:
        '201':
          description: Created risk
  /risks/summary:
    get:
      tags: [risks]
      summary: Risk summary (total, critical, high, top-5)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Risk summary
  /risks/calendar:
    get:
      tags: [risks]
      summary: Risk-related calendar events (treatment due dates)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of calendar events
  /risks/{id}:
    get:
      tags: [risks]
      summary: Get single risk
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Risk
        '404':
          description: Not found
    put:
      tags: [risks]
      summary: Update risk
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Risk'
      responses:
        '200':
          description: Updated
    delete:
      tags: [risks]
      summary: Soft-delete risk (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /risks/{id}/treatments:
    post:
      tags: [risks]
      summary: Add treatment to risk (auditor+)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                dueDate:
                  type: string
                  format: date
                status:
                  type: string
      responses:
        '201':
          description: Treatment added
  /risks/{id}/treatments/{tpId}:
    put:
      tags: [risks]
      summary: Update treatment (auditor+)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tpId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [risks]
      summary: Delete treatment (auditor+)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tpId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /risks/{id}/permanent:
    delete:
      tags: [risks]
      summary: Permanently delete risk (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permanently deleted
  /risks/{id}/restore:
    post:
      tags: [risks]
      summary: Restore soft-deleted risk (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restored

  # ─── GOALS ───────────────────────────────────────────────────────────────
...