# server/routes/risks.js
// © 2026 Claude Hecker — ISMS Builder V 1.28 — AGPL-3.0
'use strict'
const express = require('express')
const router = express.Router()
const { requireAuth, authorize } = require('../auth')
const riskStore = require('../db/riskStore')
const embeddingStore = require('../ai/embeddingStore')

function authorizeAuditor(req, res, next) {
  if (req.roleRank >= 3) return next()
  res.status(403).json({ error: 'Mindestens auditor- oder contentowner-Rolle erforderlich' })
}

router.get('/risks/calendar', requireAuth, authorize('reader'), (req, res) => {
  res.json(riskStore.getCalendarEvents())
})
router.get('/risks/summary', requireAuth, authorize('reader'), (req, res) => {
  res.json(riskStore.getSummary())
})
router.get('/risks', requireAuth, authorize('reader'), (req, res) => {
  const { category, status, entity } = req.query
  res.json(riskStore.getAll({ category, status, entity }))
})
router.get('/risks/:id', requireAuth, authorize('reader'), (req, res) => {
  const r = riskStore.getById(req.params.id)
  if (!r) return res.status(404).json({ error: 'Not found' })
  res.json(r)
})
router.post('/risks', requireAuth, authorizeAuditor, (req, res) => {
  const r = riskStore.create(req.body, req.user)
  require('../db/auditStore').append({ user: req.user, action: 'create', resource: 'risk', resourceId: r?.id, detail: req.body.title || '' })
  embeddingStore.indexDoc(r, 'Risiko', '#risks').catch(() => {})
  res.status(201).json(r)
})
router.put('/risks/:id', requireAuth, (req, res) => {
  const existing = riskStore.getById(req.params.id)
  if (!existing) return res.status(404).json({ error: 'Not found' })
  const canManageRisk = (req.roleRank || 0) >= 3
  const isOwner = existing.owner && existing.owner === req.user
  if (!canManageRisk && !isOwner) {
    return res.status(403).json({ error: 'Mindestens auditor- oder contentowner-Rolle oder eingetragener Owner erforderlich' })
  }
  const r = riskStore.update(req.params.id, req.body)
  require('../db/auditStore').append({ user: req.user, action: 'update', resource: 'risk', resourceId: req.params.id, detail: existing.title || '' })
  embeddingStore.indexDoc(r, 'Risiko', '#risks').catch(() => {})
  res.json(r)
})
router.delete('/risks/:id', requireAuth, authorize('admin'), (req, res) => {
  const ok = riskStore.delete(req.params.id, req.user)
  if (!ok) return res.status(404).json({ error: 'Not found' })
  require('../db/auditStore').append({ user: req.user, action: 'delete', resource: 'risk', resourceId: req.params.id })
  res.json({ deleted: true })
})

router.delete('/risks/:id/permanent', requireAuth, authorize('admin'), (req, res) => {
  const ok = riskStore.permanentDelete(req.params.id)
  if (!ok) return res.status(404).json({ error: 'Not found' })
  require('../db/auditStore').append({ user: req.user, action: 'permanent_delete', resource: 'risk', resourceId: req.params.id })
  embeddingStore.removeDoc(req.params.id)
  res.json({ deleted: true, permanent: true })
})

router.post('/risks/:id/restore', requireAuth, authorize('admin'), (req, res) => {
  const item = riskStore.restore(req.params.id)
  if (!item) return res.status(404).json({ error: 'Not found' })
  require('../db/auditStore').append({ user: req.user, action: 'restore', resource: 'risk', resourceId: req.params.id })
  res.json(item)
})

router.post('/risks/:id/treatments', requireAuth, authorizeAuditor, (req, res) => {
  const tp = riskStore.addTreatment(req.params.id, req.body, req.user)
  if (!tp) return res.status(404).json({ error: 'Risk not found' })
  res.status(201).json(tp)
})
router.put('/risks/:id/treatments/:tpId', requireAuth, authorizeAuditor, (req, res) => {
  const tp = riskStore.updateTreatment(req.params.id, req.params.tpId, req.body)
  if (!tp) return res.status(404).json({ error: 'Not found' })
  res.json(tp)
})
router.delete('/risks/:id/treatments/:tpId', requireAuth, authorizeAuditor, (req, res) => {
  const ok = riskStore.deleteTreatment(req.params.id, req.params.tpId)
  if (!ok) return res.status(404).json({ error: 'Not found' })
  require('../db/auditStore').append({ user: req.user, action: 'delete', resource: 'risk_treatment', resourceId: req.params.tpId, detail: `riskId: ${req.params.id}` })
  res.json({ deleted: true })
})

module.exports = router